refactor(memory): add process host seam
