+++ tests/test_memory_runtime.py
@@ -6,16 +6,10 @@ import inspect
 import json
 import logging
 import os
-import shutil
-import signal
-import stat
 import subprocess
 import sys
-import tempfile
 import threading
-import time
 from collections import deque
-from collections.abc import Iterator
 from dataclasses import replace
 from datetime import datetime, timedelta, timezone
 from pathlib import Path
@@ -35,30 +29,19 @@ from core.memory.artifact import (
 )
 from core.memory.clear_journal import MemoryClearJournal
 from core.memory.maintenance import MemoryMaintenance
-from core.memory.attachments import AttachmentPinStore, attachment_pin_root
+from core.memory.attachments import attachment_pin_root
 from core.memory.everos import (
     AddAck,
-    EverOSPort,
     FakeMemoryProvider,
     FlushSucceeded,
-    ProviderCapture,
     ProviderHealthSnapshot,
 )
-import core.memory.process as memory_process
 import core.memory.runtime as memory_runtime
 from core.memory.process import (
-    _SIDECAR_ENTRYPOINT_MODULE,
     EverOSProcess,
     EverOSProcessSettings,
     FakeEverOSProcess,
     FakeEverOSProcessFactory,
-    _ProcessIdentity,
-    _RecordedSidecar,
-    _classify_recorded_sidecar,
-    _live_owned_processes,
-    _signal_owned_group_or_process,
-    _signal_owned_processes,
-    _snapshot_owned_processes,
 )
 from core.memory.runtime import (
     MemoryRuntime,
@@ -66,9 +49,7 @@ from core.memory.runtime import (
     MemoryStoreUnavailableError,
     create_memory_runtime,
 )
-from core.memory.sidecar import _request_rejection
 from core.memory.everos_insight.recorder import initialize_call_log
-from core.memory.snapshot import MemorySnapshotManager
 from core.memory.store import MemoryStore
 from core.memory.types import (
     CaptureAccepted,
@@ -76,7 +57,6 @@ from core.memory.types import (
     MemoryItems,
     MemoryProfile,
     MemoryProfileExplicitInfo,
-    OperationFailed,
     CaptureAttachment,
     CaptureRequest,
     ProviderSessionRef,
@@ -350,71 +330,6 @@ def test_memory_runtime_reopens_the_store_after_initialization_failure(
     assert runtime.principal_for_user_key("avibe:local").startswith("u-")
 
 
-def test_refresh_owned_processes_uses_one_snapshot_and_prunes_dead_identities(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-) -> None:
-    process = EverOSProcess(sys.executable, effective_home=tmp_path, settings=_settings())
-    process._process_group = 42_424
-    process._owned_processes = {100: 1.0, 200: 2.0}
-    snapshots: list[tuple[int, int | None]] = []
-
-    def snapshot(pid: int, process_group: int | None) -> dict[int, float]:
-        snapshots.append((pid, process_group))
-        return {100: 1.0, 300: 3.0}
-
-    monkeypatch.setattr(memory_process, "_snapshot_owned_processes", snapshot)
-    monkeypatch.setattr(
-        memory_process,
-        "_live_owned_processes",
-        lambda identities: {pid: created for pid, created in identities.items() if pid != 200},
-    )
-
-    refreshed = process._refresh_owned_processes(100)
-
-    assert snapshots == [(100, 42_424)]
-    assert refreshed == {100: 1.0, 300: 3.0}
-    assert process._owned_processes == refreshed
-
-
-def test_refresh_owned_processes_retains_unverifiable_identity_sentinels(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-) -> None:
-    process = EverOSProcess(sys.executable, effective_home=tmp_path, settings=_settings())
-    process._owned_processes = {100: 1.0, 200: -1.0}
-    monkeypatch.setattr(memory_process, "_snapshot_owned_processes", lambda *_args: {100: 1.0})
-    monkeypatch.setattr(memory_process, "_live_owned_processes", lambda _identities: {100: 1.0})
-
-    assert process._refresh_owned_processes(100) == {100: 1.0, 200: -1.0}
-
-
-def test_tcp_listener_check_reuses_refreshed_owned_processes(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-) -> None:
-    process = EverOSProcess(sys.executable, effective_home=tmp_path, settings=_settings())
-    inspected: list[int] = []
-
-    class _Process:
-        def __init__(self, process_id: int) -> None:
-            self.process_id = process_id
-
-        def net_connections(self, *, kind: str):
-            assert kind == "inet"
-            inspected.append(self.process_id)
-            return []
-
-    monkeypatch.setattr(memory_process.psutil, "Process", _Process)
-    monkeypatch.setattr(
-        memory_process,
-        "_snapshot_owned_processes",
-        lambda *_args: (_ for _ in ()).throw(AssertionError("listener check repeated the process-tree snapshot")),
-    )
-
-    process._assert_no_tcp_listener(100, owned_processes={100: 1.0, 300: 3.0})
-
-    assert inspected == [100, 300]
 
 
 def test_memory_artifact_uses_shared_manager_status_shape(tmp_path: Path) -> None:
@@ -1267,200 +1182,6 @@ def test_recorded_orphan_recovery_failure_still_applies_a_disable(
     assert str(_ORPHAN_PID) in caplog.text
 
 
-def test_sidecar_child_environment_is_allowlisted_and_generated_config_has_no_keys(monkeypatch, tmp_path: Path) -> None:
-    monkeypatch.setenv("HTTP_PROXY", "http://proxy.invalid")
-    monkeypatch.setenv("SSL_CERT_FILE", "/tmp/override.pem")
-    process = EverOSProcess(
-        sys.executable,
-        effective_home=tmp_path,
-        settings=_settings(),
-    )
-    process._prepare_owned_directories()
-    process._write_generated_config()
-    environment = process._child_environment()
-    generated = (tmp_path / "memory" / "generated" / "everos.toml").read_text(encoding="utf-8")
-
-    assert environment["EVEROS_LLM__API_KEY"] == "llm-secret"
-    assert environment["EVEROS_MULTIMODAL__BASE_URL"] == environment["EVEROS_LLM__BASE_URL"]
-    assert environment["EVEROS_MULTIMODAL__MODEL"] == environment["EVEROS_LLM__MODEL"]
-    assert environment["EVEROS_MULTIMODAL__API_KEY"] == "llm-secret"
-    assert environment["AVIBE_MEMORY_ATTACHMENTS_ROOT"] == str(
-        attachment_pin_root(tmp_path)
-    )
-    assert environment["EVEROS_EMBEDDING__API_KEY"] == "embedding-secret"
-    assert "HTTP_PROXY" not in environment
-    assert "SSL_CERT_FILE" not in environment
-    assert "llm-secret" not in generated
-    assert "embedding-secret" not in generated
-    assert "rerank" in generated
-    assert str(attachment_pin_root(tmp_path)) in generated
-    assert "AVIBE_MEMORY_CALL_LOG_DB" not in environment
-
-
-def test_pinned_attachment_uri_matches_process_body_and_sidecar_guard(
-    monkeypatch, tmp_path: Path
-) -> None:
-    home = tmp_path / "avibe-home"
-    monkeypatch.setenv("AVIBE_HOME", str(home))
-    source_root = home / "attachments" / "avibe"
-    source_root.mkdir(parents=True, mode=0o700)
-    source = source_root / "diagram.png"
-    source.write_bytes(b"pinned image")
-    source.chmod(0o600)
-    pin_store = AttachmentPinStore(
-        root=attachment_pin_root(home),
-        source_root=source_root,
-    )
-    bundle = pin_store.pin(
-        (
-            CaptureAttachment(
-                kind="image",
-                name=source.name,
-                uri=source.as_uri(),
-                ext="png",
-            ),
-        )
-    )
-    pinned = pin_store.provider_attachments(bundle)
-
-    process = EverOSProcess(sys.executable, effective_home=home, settings=_settings())
-    environment = process._child_environment()
-    allowed_root = Path(environment["AVIBE_MEMORY_ATTACHMENTS_ROOT"])
-    assert allowed_root == attachment_pin_root(home)
-
-    request: dict[str, object] = {}
-
-    async def capture_request(
-        method: str,
-        route: str,
-        payload: dict[str, object],
-        *,
-        timeout_seconds: float | None = None,
-    ) -> tuple[int, bytes]:
-        request.update(method=method, route=route, payload=payload)
-        return 200, b'{"data":{"status":"accumulated"}}'
-
-    provider = EverOSPort(home / "memory" / ".rt" / "everos.sock")
-    monkeypatch.setattr(provider, "_sidecar_write", capture_request)
-    session_ref = ProviderSessionRef(
-        principal_id="u-11111111111111111111111111111111",
-        epoch=1,
-        project_ref=PROJECT,
-        session_id=f"src--{'1' * 64}--e1",
-    )
-    asyncio.run(
-        provider.add(
-            ProviderCapture(
-                session_ref=session_ref,
-                text="remember this diagram",
-                provider_timestamp_ms=1_725_000_001_234,
-                attachments=pinned,
-            )
-        )
-    )
-
-    body = json.dumps(request["payload"]).encode()
-    assert request["method"] == "POST"
-    assert request["route"] == "/api/v2/memory/add"
-    assert (
-        _request_rejection(
-            "POST",
-            "/api/v2/memory/add",
-            body,
-            attachments_root=allowed_root,
-        )
-        is None
-    )
-
-    outside = home / "outside.png"
-    outside.write_bytes(b"outside")
-    payload = request["payload"]
-    assert isinstance(payload, dict)
-    messages = payload["messages"]
-    assert isinstance(messages, list)
-    content = messages[0]["content"]
-    assert isinstance(content, list)
-    content[-1]["uri"] = outside.as_uri()
-    assert (
-        _request_rejection(
-            "POST",
-            "/api/v2/memory/add",
-            json.dumps(payload).encode(),
-            attachments_root=allowed_root,
-        )
-        == "add"
-    )
-
-
-async def test_runtime_effective_home_owns_the_attachment_pipeline(
-    monkeypatch, tmp_path: Path
-) -> None:
-    global_home = tmp_path / "global-home"
-    runtime_home = tmp_path / "runtime-home"
-    monkeypatch.setenv("AVIBE_HOME", str(runtime_home))
-    store = MemoryStore()
-    monkeypatch.setenv("AVIBE_HOME", str(global_home))
-    source_root = runtime_home / "attachments" / "avibe"
-    source_root.mkdir(parents=True, mode=0o700)
-    source = source_root / "notes.txt"
-    source.write_bytes(b"runtime-owned attachment")
-    source.chmod(0o600)
-
-    runtime = MemoryRuntime(
-        MemoryConfig(enabled=True),
-        store=store,
-        artifact_manager=_installed_artifact(),
-        effective_home=runtime_home,
-    )
-    expected_pin_root = attachment_pin_root(runtime_home)
-    assert runtime.module._effective_home == runtime_home
-    assert runtime.module._attachment_store._effective_home == runtime_home
-    assert runtime.module._attachment_store._root == expected_pin_root
-    assert runtime.module._attachment_store._source_root == source_root
-    manager = _maintenance(runtime)._snapshot_manager
-    assert manager is not None
-    assert manager.effective_home == runtime_home
-    assert any(
-        surface.path == "memory/attachments"
-        for surface in manager.surfaces
-    )
-
-    process = EverOSProcess(
-        sys.executable,
-        effective_home=runtime_home,
-        settings=_settings(),
-    )
-    assert process._child_environment()["AVIBE_MEMORY_ATTACHMENTS_ROOT"] == str(
-        expected_pin_root
-    )
-
-    try:
-        result = await runtime.module.capture(
-            CaptureRequest(
-                source_message_id="source-1",
-                session_id="session-1",
-                principal_id="u-11111111111111111111111111111111",
-                project_id=PROJECT,
-                provenance="user_input",
-                text="remember the notes",
-                occurred_at_ms=1_725_000_001_234,
-                attachments=(
-                    CaptureAttachment(
-                        kind="doc",
-                        name=source.name,
-                        uri=source.as_uri(),
-                        ext="txt",
-                    ),
-                ),
-            )
-        )
-    finally:
-        await runtime.close()
-    assert isinstance(result, CaptureAccepted)
-    pinned_files = tuple((expected_pin_root / "bundles").glob("*/*"))
-    assert len(pinned_files) == 1
-    assert pinned_files[0].read_bytes() == b"runtime-owned attachment"
-    assert not attachment_pin_root(global_home).exists()
 
 
 async def test_final_flush_fences_capture_before_queue_visibility(
@@ -1914,396 +1635,6 @@ async def test_final_flush_deadline_includes_capture_admission_wait(
         await runtime.close()
 
 
-def test_sidecar_child_environment_includes_only_the_configured_call_log(tmp_path: Path) -> None:
-    call_log = tmp_path / "memory" / "call-log" / "call-log.db"
-    process = EverOSProcess(
-        sys.executable,
-        effective_home=tmp_path,
-        settings=replace(_settings(), call_log_db_path=call_log),
-    )
-
-    process._prepare_owned_directories()
-    environment = process._child_environment()
-
-    assert environment["AVIBE_MEMORY_CALL_LOG_DB"] == str(call_log)
-    assert stat.S_IMODE(call_log.parent.stat().st_mode) == 0o700
-
-
-async def test_sidecar_rejects_sun_path_overflow_without_launching_child(tmp_path: Path) -> None:
-    socket_path = tmp_path / ("a" * 180) / "everos.sock"
-    process = EverOSProcess(
-        sys.executable,
-        effective_home=tmp_path,
-        socket_path=socket_path,
-        settings=_settings(),
-    )
-
-    assert await process.start() is False
-    assert process.last_error == "memory_sidecar_unavailable"
-    assert process.consecutive_failures == 1
-    await process.stop()
-
-
-async def test_sidecar_start_failure_never_relaunches_beside_an_unreaped_child(monkeypatch, tmp_path: Path) -> None:
-    class _Child:
-        pid = 999_999
-        returncode = None
-
-        async def wait(self) -> None:
-            return None
-
-        def send_signal(self, _signum) -> None:
-            return None
-
-    launches: list[_Child] = []
-
-    async def spawn(*_args, **_kwargs) -> _Child:
-        child = _Child()
-        launches.append(child)
-        return child
-
-    async def readiness_failure(_process) -> None:
-        raise RuntimeError("readiness failed")
-
-    async def cleanup_failure(*_args, **_kwargs) -> None:
-        raise RuntimeError("child tree still alive")
-
-    process = EverOSProcess(
-        sys.executable,
-        effective_home=tmp_path,
-        socket_path=Path(f"/tmp/everos-{os.getpid()}.sock"),
-        settings=_settings(),
-    )
-    monkeypatch.setattr("core.memory.process.asyncio.create_subprocess_exec", spawn)
-    monkeypatch.setattr(process, "_prepare_owned_directories", lambda: None)
-    monkeypatch.setattr(process, "_write_generated_config", lambda: None)
-    monkeypatch.setattr(process, "_remove_owned_socket", lambda: None)
-    monkeypatch.setattr(process, "_wait_for_ready", readiness_failure)
-    monkeypatch.setattr(process, "_terminate_owned_tree", cleanup_failure)
-
-    assert await process.start() is False
-    assert process.down is True
-    assert process._process is launches[0]
-    assert process._restart_task is None
-    assert await process.start() is False
-    assert len(launches) == 1
-
-
-async def test_processing_probe_reaps_child_when_its_caller_is_cancelled(monkeypatch, tmp_path: Path) -> None:
-    started = asyncio.Event()
-    cleanup_calls: list[object] = []
-
-    class _Probe:
-        pid = 999_999
-        returncode = None
-
-        async def wait(self) -> None:
-            started.set()
-            await asyncio.Event().wait()
-
-    async def spawn(*_args, **_kwargs) -> _Probe:
-        return _Probe()
-
-    async def cleanup(*_args, **_kwargs) -> None:
-        cleanup_calls.append(object())
-
-    process = EverOSProcess(
-        sys.executable,
-        effective_home=tmp_path,
-        settings=_settings(),
-    )
-    monkeypatch.setattr("core.memory.process.asyncio.create_subprocess_exec", spawn)
-    monkeypatch.setattr(process, "_terminate_owned_tree", cleanup)
-
-    task = asyncio.create_task(process.processing_healthy())
-    await started.wait()
-    task.cancel()
-    try:
-        await task
-    except asyncio.CancelledError:
-        pass
-    else:
-        raise AssertionError("processing probe cancellation was swallowed")
-    assert cleanup_calls
-
-
-async def test_sidecar_stop_signals_isolated_child_group(tmp_path: Path) -> None:
-    child_pid_path = tmp_path / "child.pid"
-    script = (
-        "import pathlib, subprocess, sys, time; "
-        f"child = subprocess.Popen([sys.executable, '-c', 'import time; time.sleep(60)']); "
-        f"pathlib.Path({str(child_pid_path)!r}).write_text(str(child.pid)); time.sleep(60)"
-    )
-
-    child = await asyncio.create_subprocess_exec(sys.executable, "-c", script, start_new_session=True)
-    deadline = time.monotonic() + 3
-    while not child_pid_path.exists() and time.monotonic() < deadline:
-        await asyncio.sleep(0.02)
-    assert child_pid_path.exists()
-    descendant_pid = int(child_pid_path.read_text(encoding="utf-8"))
-    process = EverOSProcess(sys.executable, effective_home=tmp_path, settings=_settings())
-    process._process_group = os.getpgid(child.pid)
-    owned_processes = _snapshot_owned_processes(child.pid, process._process_group)
-    await process._terminate_owned_tree(
-        child,
-        process_group=process._process_group,
-        owned_processes=owned_processes,
-    )
-    parent_pid = child.pid
-    assert not _pid_exists(parent_pid)
-    assert not _pid_exists(descendant_pid)
-
-
-def test_sidecar_safety_monitor_ignores_expected_shutdown(monkeypatch, tmp_path: Path, caplog) -> None:
-    class _Child:
-        pid = 999_999
-        returncode = None
-
-    process = EverOSProcess(sys.executable, effective_home=tmp_path, settings=_settings())
-    child = _Child()
-    process._process = child
-    process._desired_running = False
-    monkeypatch.setattr(memory_process, "_owned_process_identity_is_live", lambda *_args: False)
-
-    asyncio.run(process._monitor_child(child))
-
-    assert "safety monitor rejected" not in caplog.text
-    assert process._process is child
-
-
-async def test_sidecar_stop_reaps_a_descendant_that_leaves_the_child_group(tmp_path: Path) -> None:
-    child_pid_path = tmp_path / "detached-child.pid"
-    script = (
-        "import pathlib, subprocess, sys, time; "
-        f"child = subprocess.Popen([sys.executable, '-c', 'import time; time.sleep(60)'], start_new_session=True); "
-        f"pathlib.Path({str(child_pid_path)!r}).write_text(str(child.pid)); time.sleep(60)"
-    )
-
-    child = await asyncio.create_subprocess_exec(sys.executable, "-c", script, start_new_session=True)
-    deadline = time.monotonic() + 3
-    while not child_pid_path.exists() and time.monotonic() < deadline:
-        await asyncio.sleep(0.02)
-    assert child_pid_path.exists()
-    descendant_pid = int(child_pid_path.read_text(encoding="utf-8"))
-    process = EverOSProcess(sys.executable, effective_home=tmp_path, settings=_settings())
-    process._process_group = os.getpgid(child.pid)
-    owned_processes = _snapshot_owned_processes(child.pid, process._process_group)
-    await process._terminate_owned_tree(
-        child,
-        process_group=process._process_group,
-        owned_processes=owned_processes,
-    )
-    parent_pid = child.pid
-    assert not _pid_exists(parent_pid)
-    assert not _pid_exists(descendant_pid)
-
-
-def test_sidecar_cleanup_skips_a_reused_pid_identity(monkeypatch, tmp_path: Path) -> None:
-    signals: list[int] = []
-
-    class _Child:
-        pid = 42_424
-        returncode = None
-
-        async def wait(self) -> None:
-            return None
-
-        def send_signal(self, signum: int) -> None:
-            signals.append(signum)
-
-    async def reaped(*_args, **_kwargs) -> bool:
-        return True
-
-    process = EverOSProcess(sys.executable, effective_home=tmp_path, settings=_settings())
-    monkeypatch.setattr("core.memory.process._snapshot_owned_processes", lambda *_args: {42_424: 22.0})
-    monkeypatch.setattr("core.memory.process._wait_for_owned_exit", reaped)
-
-    asyncio.run(
-        process._terminate_owned_tree(
-            _Child(),
-            process_group=None,
-            owned_processes={42_424: 11.0},
-        )
-    )
-
-    assert signals == []
-
-
-async def test_sidecar_cleanup_never_signals_spawned_pid_after_identity_changes(monkeypatch, tmp_path: Path) -> None:
-    signals: list[tuple[str, int]] = []
-
-    class _TrackedChild:
-        returncode = None
-
-        def __init__(self, pid: int) -> None:
-            self.pid = pid
-
-        def send_signal(self, signum: int) -> None:
-            signals.append(("child", signum))
-
-    child = await asyncio.create_subprocess_exec(
-        sys.executable,
-        "-c",
-        "import time; time.sleep(60)",
-        start_new_session=True,
-    )
-    try:
-        identities = _snapshot_owned_processes(child.pid, None)
-        captured_at = identities[child.pid]
-        original_process = memory_process.psutil.Process
-
-        class _ReusedProcess:
-            def __init__(self, process_id: int) -> None:
-                assert process_id == child.pid
-                self.pid = process_id
-
-            def create_time(self) -> float:
-                return captured_at + 1.0
-
-            def status(self) -> str:
-                return psutil.STATUS_SLEEPING
-
-            def send_signal(self, signum: int) -> None:
-                signals.append(("psutil", signum))
-
-        monkeypatch.setattr(memory_process.psutil, "Process", _ReusedProcess)
-        try:
-            _signal_owned_processes(identities, signal.SIGTERM)
-            _signal_owned_group_or_process(_TrackedChild(child.pid), None, identities, signal.SIGTERM)
-        finally:
-            monkeypatch.setattr(memory_process.psutil, "Process", original_process)
-
-        assert signals == []
-    finally:
-        if child.returncode is None:
-            child.terminate()
-            try:
-                await asyncio.wait_for(child.wait(), timeout=3.0)
-            except TimeoutError:
-                child.kill()
-                await child.wait()
-
-
-def test_sidecar_cleanup_does_not_group_signal_an_unconfirmed_member(monkeypatch) -> None:
-    group_signals: list[tuple[int, int]] = []
-    child_signals: list[int] = []
-
-    class _TrackedChild:
-        pid = 42_424
-        returncode = None
-
-        def send_signal(self, signum: int) -> None:
-            child_signals.append(signum)
-
-    monkeypatch.setattr(memory_process, "_snapshot_process_group", lambda _group: {42_424: 11.0, 42_425: 12.0})
-    monkeypatch.setattr(memory_process, "_confirmed_owned_processes", lambda _identities: {42_424: 11.0})
-    monkeypatch.setattr(memory_process.os, "killpg", lambda group, signum: group_signals.append((group, signum)))
-
-    _signal_owned_group_or_process(
-        _TrackedChild(),
-        42_424,
-        {42_424: 11.0, 42_425: 12.0},
-        signal.SIGTERM,
-    )
-
-    assert group_signals == []
-    assert child_signals == [signal.SIGTERM]
-
-
-def test_sidecar_group_snapshot_fails_closed_for_an_inaccessible_member(monkeypatch) -> None:
-    parent_id = 42_424
-    child_id = 42_425
-    group_id = 42_424
-    group_signals: list[tuple[int, int]] = []
-    child_signals: list[int] = []
-
-    class _GroupMember:
-        def __init__(self, process_id: int, created_at: float | None) -> None:
-            self.pid = process_id
-            self._created_at = created_at
-
-        def create_time(self) -> float:
-            if self._created_at is None:
-                raise psutil.AccessDenied(pid=self.pid)
-            return self._created_at
-
-    class _TrackedChild:
-        pid = parent_id
-        returncode = None
-
-        def send_signal(self, signum: int) -> None:
-            child_signals.append(signum)
-
-    monkeypatch.setattr(
-        memory_process.psutil,
-        "process_iter",
-        lambda: [_GroupMember(parent_id, 11.0), _GroupMember(child_id, None)],
-    )
-    monkeypatch.setattr(memory_process.os, "getpgid", lambda _process_id: group_id)
-    monkeypatch.setattr(memory_process, "_confirmed_owned_processes", lambda _identities: {parent_id: 11.0})
-    monkeypatch.setattr(memory_process.os, "killpg", lambda group, signum: group_signals.append((group, signum)))
-
-    snapshot = memory_process._snapshot_process_group(group_id)
-    declared_safe = memory_process._group_contains_only_confirmed_owned_processes(
-        group_id,
-        {parent_id: 11.0, child_id: 12.0},
-    )
-    _signal_owned_group_or_process(
-        _TrackedChild(),
-        group_id,
-        {parent_id: 11.0, child_id: 12.0},
-        signal.SIGTERM,
-    )
-
-    assert snapshot == {parent_id: 11.0, child_id: -1.0}
-    assert declared_safe is False
-    assert group_signals == []
-    assert child_signals == [signal.SIGTERM]
-
-
-def test_sidecar_cleanup_keeps_access_denied_identity_live_without_signaling(monkeypatch) -> None:
-    process_id = 42_425
-
-    class _InaccessibleProcess:
-        def __init__(self, _process_id: int) -> None:
-            raise psutil.AccessDenied(pid=process_id)
-
-    monkeypatch.setattr(memory_process.psutil, "Process", _InaccessibleProcess)
-    identities = {process_id: 11.0}
-
-    assert _live_owned_processes(identities) == identities
-    _signal_owned_processes(identities, signal.SIGTERM)
-
-
-def test_sidecar_crash_counter_resets_only_after_observed_healthy_window(tmp_path: Path) -> None:
-    process = EverOSProcess(sys.executable, effective_home=tmp_path, settings=_settings())
-    process._consecutive_failures = 4
-
-    process._record_health_observation(True, observed_at=10.0)
-    process._record_health_observation(False, observed_at=310.0)
-    process._record_health_observation(True, observed_at=311.0)
-    process._record_health_observation(True, observed_at=610.0)
-
-    assert process.consecutive_failures == 4
-
-    process._record_health_observation(True, observed_at=611.0)
-
-    assert process.consecutive_failures == 0
-
-
-def test_explicit_sidecar_retry_keeps_crash_budget_until_observed_health(monkeypatch, tmp_path: Path) -> None:
-    process = EverOSProcess(sys.executable, effective_home=tmp_path, settings=_settings())
-    process._down = True
-    process._consecutive_failures = 5
-
-    async def start_stub() -> bool:
-        return True
-
-    monkeypatch.setattr(process, "_start_locked", start_stub)
-
-    assert asyncio.run(process.start()) is True
-    assert process.down is False
-    assert process.consecutive_failures == 5
 
 
 async def test_runtime_exposes_interrupted_clear_without_starting_sidecar(tmp_path: Path) -> None:
@@ -3176,22 +2507,6 @@ async def test_clear_quiesce_does_not_delete_call_log_files(
     await runtime.close()
 
 
-def test_generated_timezone_stays_with_existing_provider_root(tmp_path: Path) -> None:
-    process = EverOSProcess(
-        sys.executable,
-        effective_home=tmp_path,
-        settings=_settings(),
-    )
-    process._prepare_owned_directories()
-    (tmp_path / "memory" / "everos-root" / "everos.toml").write_text(
-        "[memory]\ntimezone = \"Asia/Shanghai\"\n",
-        encoding="utf-8",
-    )
-
-    process._write_generated_config()
-
-    contents = (tmp_path / "memory" / "everos-root" / "everos.toml").read_text(encoding="utf-8")
-    assert 'timezone = "Asia/Shanghai"' in contents
 
 
 def _pid_exists(pid: int) -> bool:
@@ -4425,1348 +3740,72 @@ _FOREIGN_UID_GROUP_PID = 424_247
 _ORPHAN_CREATE_TIME = 1_700_000_000.5
 
 
-def _orphan_process(tmp_path: Path, **overrides) -> EverOSProcess:
-    return EverOSProcess(
-        sys.executable,
-        effective_home=tmp_path,
-        settings=_settings(),
-        **overrides,
+async def test_runtime_effective_home_owns_the_attachment_pipeline(
+    monkeypatch, tmp_path: Path
+) -> None:
+    global_home = tmp_path / "global-home"
+    runtime_home = tmp_path / "runtime-home"
+    monkeypatch.setenv("AVIBE_HOME", str(runtime_home))
+    store = MemoryStore()
+    monkeypatch.setenv("AVIBE_HOME", str(global_home))
+    source_root = runtime_home / "attachments" / "avibe"
+    source_root.mkdir(parents=True, mode=0o700)
+    source = source_root / "notes.txt"
+    source.write_bytes(b"runtime-owned attachment")
+    source.chmod(0o600)
+
+    runtime = MemoryRuntime(
+        MemoryConfig(enabled=True),
+        store=store,
+        artifact_manager=_installed_artifact(),
+        effective_home=runtime_home,
+    )
+    expected_pin_root = attachment_pin_root(runtime_home)
+    assert runtime.module._effective_home == runtime_home
+    assert runtime.module._attachment_store._effective_home == runtime_home
+    assert runtime.module._attachment_store._root == expected_pin_root
+    assert runtime.module._attachment_store._source_root == source_root
+    manager = _maintenance(runtime)._snapshot_manager
+    assert manager is not None
+    assert manager.effective_home == runtime_home
+    assert any(
+        surface.path == "memory/attachments"
+        for surface in manager.surfaces
     )
 
+    process = EverOSProcess(
+        sys.executable,
+        effective_home=runtime_home,
+        settings=_settings(),
+    )
+    assert process._child_environment()["AVIBE_MEMORY_ATTACHMENTS_ROOT"] == str(
+        expected_pin_root
+    )
 
-@pytest.fixture
-def short_socket_path() -> Iterator[Path]:
-    """A socket path that fits ``sun_path``, for tests that call ``start``.
-
-    ``tmp_path`` alone is already past the 104-byte macOS limit, so a launch rooted
-    there fails ``_validate_launch_inputs`` before it ever reaches the orphan
-    check — every assertion after it would hold vacuously.
-    """
-
-    directory = Path(tempfile.mkdtemp(prefix="avibe-"))
-    socket_path = directory / "everos.sock"
-    assert len(os.fsencode(socket_path)) + 1 <= 104, socket_path
     try:
-        yield socket_path
+        result = await runtime.module.capture(
+            CaptureRequest(
+                source_message_id="source-1",
+                session_id="session-1",
+                principal_id="u-11111111111111111111111111111111",
+                project_id=PROJECT,
+                provenance="user_input",
+                text="remember the notes",
+                occurred_at_ms=1_725_000_001_234,
+                attachments=(
+                    CaptureAttachment(
+                        kind="doc",
+                        name=source.name,
+                        uri=source.as_uri(),
+                        ext="txt",
+                    ),
+                ),
+            )
+        )
     finally:
-        shutil.rmtree(directory, ignore_errors=True)
-
-
-def _orphan_record(process: EverOSProcess, **overrides) -> dict:
-    record = {
-        "pid": _ORPHAN_PID,
-        "create_time": _ORPHAN_CREATE_TIME,
-        # ``start_new_session=True`` makes the sidecar lead a group of its own
-        # number, which is what identifies its helpers once the leader is gone.
-        "process_group": _ORPHAN_PID,
-        "socket_path": str(process.socket_path),
-        "provider_root": str(process.provider_root),
-    }
-    record.update(overrides)
-    return record
-
-
-def _orphan_identity(process: EverOSProcess, **overrides) -> _ProcessIdentity:
-    fields = {
-        "create_time": _ORPHAN_CREATE_TIME,
-        "cmdline": (
-            sys.executable,
-            "-m",
-            _SIDECAR_ENTRYPOINT_MODULE,
-            "--uds",
-            str(process.socket_path),
-        ),
-        "uid": os.getuid() if hasattr(os, "getuid") else None,
-    }
-    fields.update(overrides)
-    return _ProcessIdentity(**fields)
-
-
-def _write_orphan_record(process: EverOSProcess, record: dict) -> Path:
-    path = process._ownership.record_path
-    path.parent.mkdir(parents=True, exist_ok=True)
-    path.write_text(json.dumps(record), encoding="utf-8")
-    return path
-
-
-def test_recorded_sidecar_identity_accepts_only_a_provably_owned_orphan(tmp_path: Path) -> None:
-    """The decision that gates a kill signal, in one place.
-
-    Every ``NOT_OURS`` case below is a process Avibe must leave alone and a record
-    it may retire: a recycled pid, a sidecar from another home, another user's
-    process, or something that is not our entrypoint at all. ``UNVERIFIABLE`` is
-    the separate case of a live pid whose deciding facts were never disclosed --
-    it must not be confused with "gone", because that starts a second sidecar.
-    """
-
-    process = _orphan_process(tmp_path)
-    own_uid = os.getuid() if hasattr(os, "getuid") else None
-
-    def verdict(record: dict, identity: _ProcessIdentity | None) -> _RecordedSidecar:
-        return _classify_recorded_sidecar(
-            record,
-            identity,
-            socket_path=process.socket_path,
-            provider_root=process.provider_root,
-        )
-
-    assert verdict(_orphan_record(process), _orphan_identity(process)) is _RecordedSidecar.OURS
-
-    not_ours: list[tuple[dict, _ProcessIdentity | None]] = [
-        # The process is confirmed gone.
-        (_orphan_record(process), None),
-        # The pid was recycled: same number, different process.
-        (_orphan_record(process), _orphan_identity(process, create_time=_ORPHAN_CREATE_TIME + 1)),
-        # Not our entrypoint.
-        (_orphan_record(process), _orphan_identity(process, cmdline=(sys.executable, "-m", "http.server"))),
-        # Our entrypoint name, but serving a different socket.
-        (
-            _orphan_record(process),
-            _orphan_identity(
-                process,
-                cmdline=(sys.executable, "-m", _SIDECAR_ENTRYPOINT_MODULE, "--uds", "/tmp/other.sock"),
-            ),
-        ),
-        # Another user's process.
-        (_orphan_record(process), _orphan_identity(process, uid=(own_uid or 0) + 1)),
-        # A recycled pid owned by another user that will not disclose its cmdline:
-        # the readable uid alone is enough to rule it out, so startup continues.
-        (
-            _orphan_record(process),
-            _orphan_identity(process, uid=(own_uid or 0) + 1, cmdline=None),
-        ),
-        # A record written for a different provider root or socket.
-        (_orphan_record(process, provider_root="/tmp/other-root"), _orphan_identity(process)),
-        (_orphan_record(process, socket_path="/tmp/other.sock"), _orphan_identity(process)),
-        # A malformed creation time can never be matched.
-        (_orphan_record(process, create_time="1700000000.5"), _orphan_identity(process)),
-        (_orphan_record(process, create_time=True), _orphan_identity(process)),
-    ]
-    for record, identity in not_ours:
-        assert verdict(record, identity) is _RecordedSidecar.NOT_OURS, (record, identity)
-
-    unverifiable: list[tuple[dict, _ProcessIdentity]] = [
-        # Live, matching uid and creation time, but the cmdline is withheld —
-        # nothing here excludes our own sidecar.
-        (_orphan_record(process), _orphan_identity(process, cmdline=None)),
-        # A live pid that disclosed nothing at all.
-        (_orphan_record(process), _orphan_identity(process, create_time=None, cmdline=None, uid=None)),
-        # The creation time alone is withheld, so pid reuse cannot be ruled out.
-        (_orphan_record(process), _orphan_identity(process, create_time=None)),
-    ]
-    for record, identity in unverifiable:
-        assert verdict(record, identity) is _RecordedSidecar.UNVERIFIABLE, (record, identity)
-
-    if own_uid is not None:
-        # An unreadable uid is not an exclusion either.
-        assert verdict(_orphan_record(process), _orphan_identity(process, uid=None)) is (
-            _RecordedSidecar.UNVERIFIABLE
-        )
-
-
-def _guarded_process_class(
-    *,
-    create_time: float | None = _ORPHAN_CREATE_TIME,
-    uid: int | None = None,
-    cmdline: tuple[str, ...] | None = None,
-):
-    """A ``psutil.Process`` stand-in that withholds every field left at ``None``.
-
-    Models what a real OS does: macOS discloses ``create_time`` and ``uids`` for
-    any pid but refuses ``cmdline`` outside the caller's own uid.
-    """
-
-    class _Guarded:
-        def __init__(self, process_id: int) -> None:
-            self.pid = process_id
-
-        def status(self) -> str:
-            return psutil.STATUS_SLEEPING
-
-        def create_time(self) -> float:
-            if create_time is None:
-                raise psutil.AccessDenied(pid=self.pid)
-            return create_time
-
-        def cmdline(self) -> list[str]:
-            if cmdline is None:
-                raise psutil.AccessDenied(pid=self.pid)
-            return list(cmdline)
-
-        def uids(self):
-            if uid is None:
-                raise psutil.AccessDenied(pid=self.pid)
-            return SimpleNamespace(real=uid, effective=uid, saved=uid)
-
-    return _Guarded
-
-
-def test_process_identity_reports_undisclosed_fields_instead_of_gone(
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """A withheld field must not read as "this pid is not running".
-
-    Collapsing every ``psutil.Error`` to ``None`` made a live sidecar whose
-    cmdline the OS refuses to disclose indistinguishable from a reaped one.
-    """
-
-    guarded = _guarded_process_class(uid=4_242)
-    monkeypatch.setattr(memory_process.psutil, "Process", guarded)
-
-    identity = memory_process._inspect_process_identity(_ORPHAN_PID)
-
-    assert identity == _ProcessIdentity(create_time=_ORPHAN_CREATE_TIME, cmdline=None, uid=4_242)
-
-    class _Zombie(guarded):
-        def status(self) -> str:
-            return psutil.STATUS_ZOMBIE
-
-    class _Gone(guarded):
-        def __init__(self, process_id: int) -> None:
-            raise psutil.NoSuchProcess(pid=process_id)
-
-    class _ExitsMidRead(guarded):
-        def cmdline(self) -> list[str]:
-            raise psutil.NoSuchProcess(pid=self.pid)
-
-    for stub in (_Zombie, _Gone, _ExitsMidRead):
-        monkeypatch.setattr(memory_process.psutil, "Process", stub)
-        assert memory_process._inspect_process_identity(_ORPHAN_PID) is None
-
-    class _NoUidsPlatform(guarded):
-        def uids(self):
-            # ``psutil`` declares ``uids`` everywhere but only implements it on
-            # POSIX, so on Windows the call itself raises.
-            raise AttributeError("uids")
-
-    monkeypatch.setattr(memory_process.psutil, "Process", _NoUidsPlatform)
-
-    assert memory_process._inspect_process_identity(_ORPHAN_PID) == _ProcessIdentity(
-        create_time=_ORPHAN_CREATE_TIME,
-        cmdline=None,
-        uid=None,
-    )
-
-
-def test_sidecar_launch_reaps_a_recorded_orphan_from_a_previous_run(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-) -> None:
-    """A crashed service leaves its ``start_new_session`` child running.
-
-    Boot used to spawn a second sidecar beside it, so the orphan kept serving the
-    socket and holding handles on provider data until it was killed by hand.
-    """
-
-    process = _orphan_process(tmp_path)
-    record_path = _write_orphan_record(process, _orphan_record(process))
-    live = {_ORPHAN_PID: _ORPHAN_CREATE_TIME}
-    signalled: list[int] = []
-
-    def signal_processes(identities, signum) -> None:
-        signalled.append(signum)
-        assert identities == {_ORPHAN_PID: _ORPHAN_CREATE_TIME}
-        live.clear()
-
-    monkeypatch.setattr(
-        memory_process,
-        "_inspect_process_identity",
-        lambda pid: _orphan_identity(process) if pid == _ORPHAN_PID else None,
-    )
-    monkeypatch.setattr(memory_process, "_signal_owned_processes", signal_processes)
-    monkeypatch.setattr(memory_process, "_live_owned_processes", lambda _identities: dict(live))
-
-    asyncio.run(process._ownership.reap())
-
-    assert signalled == [signal.SIGTERM]
-    assert not record_path.exists()
-
-
-def test_sidecar_launch_never_signals_a_pid_it_cannot_identify(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-) -> None:
-    """A recycled pid retires the record instead of killing its new owner."""
-
-    process = _orphan_process(tmp_path)
-    record_path = _write_orphan_record(process, _orphan_record(process))
-    signalled: list[int] = []
-
-    monkeypatch.setattr(
-        memory_process,
-        "_inspect_process_identity",
-        lambda _pid: _orphan_identity(process, create_time=_ORPHAN_CREATE_TIME + 10),
-    )
-    monkeypatch.setattr(
-        memory_process,
-        "_signal_owned_processes",
-        lambda *_args: signalled.append(object()),
-    )
-
-    asyncio.run(process._ownership.reap())
-
-    assert signalled == []
-    assert not record_path.exists()
-
-
-def test_sidecar_launch_refuses_to_start_beside_an_unreapable_orphan(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-    short_socket_path: Path,
-) -> None:
-    """Fail closed, exactly as ``start`` already does for an unreaped child."""
-
-    process = _orphan_process(tmp_path, stop_timeout_seconds=0.1, socket_path=short_socket_path)
-    _write_orphan_record(process, _orphan_record(process))
-    spawns: list[tuple] = []
-
-    async def spawn(*args, **_kwargs):
-        # Recorded rather than asserted: ``_start_locked`` catches every exception,
-        # so a raised ``AssertionError`` here would be swallowed into a plain
-        # "start failed" and the test would pass for the wrong reason.
-        spawns.append(args)
-        raise OSError("stop before a real sidecar is spawned")
-
-    monkeypatch.setattr(
-        memory_process,
-        "_inspect_process_identity",
-        lambda _pid: _orphan_identity(process),
-    )
-    monkeypatch.setattr(memory_process, "_signal_owned_processes", lambda *_args: None)
-    monkeypatch.setattr(memory_process, "_live_owned_processes", lambda identities: dict(identities))
-    monkeypatch.setattr("core.memory.process.asyncio.create_subprocess_exec", spawn)
-
-    assert asyncio.run(process.start()) is False
-    assert process.last_error == "memory_sidecar_unavailable"
-    assert spawns == []
-    assert process._ownership.record_path.exists()
-
-
-def test_sidecar_launch_reaps_the_whole_orphan_tree_not_just_the_recorded_pid(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-) -> None:
-    """An orphan's helpers hold the provider root just as its root process does.
-
-    Signalling only the recorded pid left same-group helpers running against the
-    root while a replacement sidecar started, recreating the overlap this reap
-    exists to prevent. Discovery must match the normal stop path: descendants plus
-    every member of the isolated process group.
-    """
-
-    process = _orphan_process(tmp_path)
-    record_path = _write_orphan_record(process, _orphan_record(process))
-    tree = {
-        _ORPHAN_PID: _ORPHAN_CREATE_TIME,
-        _ORPHAN_DESCENDANT_PID: _ORPHAN_CREATE_TIME + 1,
-        _ORPHAN_GROUP_MEMBER_PID: _ORPHAN_CREATE_TIME + 2,
-    }
-    live = dict(tree)
-    snapshots: list[tuple[int, int | None]] = []
-    group_signals: list[tuple[int, int]] = []
-    signalled: list[dict[int, float]] = []
-
-    def snapshot(pid: int, process_group: int | None) -> dict[int, float]:
-        snapshots.append((pid, process_group))
-        return dict(tree)
-
-    def signal_processes(identities, signum) -> None:
-        del signum
-        signalled.append(dict(identities))
-        live.clear()
-
-    monkeypatch.setattr(
-        memory_process,
-        "_inspect_process_identity",
-        lambda pid: _orphan_identity(process) if pid == _ORPHAN_PID else None,
-    )
-    # ``start_new_session=True`` makes the sidecar its own process group leader.
-    monkeypatch.setattr(memory_process, "_isolated_process_group", lambda pid: pid)
-    monkeypatch.setattr(memory_process, "_snapshot_owned_processes", snapshot)
-    monkeypatch.setattr(memory_process, "_snapshot_process_group", lambda _group: dict(tree))
-    monkeypatch.setattr(memory_process, "_confirmed_owned_processes", lambda identities: dict(identities))
-    monkeypatch.setattr(memory_process.os, "killpg", lambda group, signum: group_signals.append((group, signum)))
-    monkeypatch.setattr(memory_process, "_signal_owned_processes", signal_processes)
-    monkeypatch.setattr(memory_process, "_live_owned_processes", lambda _identities: dict(live))
-
-    asyncio.run(process._ownership.reap())
-
-    assert snapshots == [(_ORPHAN_PID, _ORPHAN_PID)]
-    assert group_signals == [(_ORPHAN_PID, signal.SIGTERM)]
-    assert signalled == [tree]
-    assert not record_path.exists()
-
-
-def test_sidecar_orphan_reap_refuses_a_group_signal_for_an_unverifiable_member(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-) -> None:
-    """Widening discovery must not widen the blast radius.
-
-    A group holding a member carrying the ``AccessDenied`` sentinel is never
-    signaled group-wide, and a member that cannot be proven reaped still fails the
-    launch instead of being written off.
-    """
-
-    process = _orphan_process(tmp_path, stop_timeout_seconds=0.1)
-    record_path = _write_orphan_record(process, _orphan_record(process))
-    discovered = {_ORPHAN_PID: _ORPHAN_CREATE_TIME, _ORPHAN_GROUP_MEMBER_PID: -1.0}
-    live = dict(discovered)
-    group_signals: list[tuple[int, int]] = []
-    signalled: list[int] = []
-
-    def signal_processes(identities, signum) -> None:
-        del identities
-        signalled.append(signum)
-        # The confirmed root exits; the unverifiable member cannot be proven gone.
-        live.pop(_ORPHAN_PID, None)
-
-    monkeypatch.setattr(
-        memory_process,
-        "_inspect_process_identity",
-        lambda _pid: _orphan_identity(process),
-    )
-    monkeypatch.setattr(memory_process, "_isolated_process_group", lambda pid: pid)
-    monkeypatch.setattr(memory_process, "_snapshot_owned_processes", lambda _pid, _group: dict(discovered))
-    monkeypatch.setattr(memory_process, "_snapshot_process_group", lambda _group: dict(discovered))
-    monkeypatch.setattr(
-        memory_process,
-        "_confirmed_owned_processes",
-        lambda identities: {pid: created_at for pid, created_at in identities.items() if created_at >= 0},
-    )
-    monkeypatch.setattr(memory_process.os, "killpg", lambda group, signum: group_signals.append((group, signum)))
-    monkeypatch.setattr(memory_process, "_signal_owned_processes", signal_processes)
-    monkeypatch.setattr(memory_process, "_live_owned_processes", lambda _identities: dict(live))
-
-    with pytest.raises(RuntimeError, match="orphaned sidecar did not exit"):
-        asyncio.run(process._ownership.reap())
-
-    assert group_signals == []
-    assert signalled == [signal.SIGTERM, getattr(signal, "SIGKILL", signal.SIGTERM)]
-    assert record_path.exists()
-
-
-def test_sidecar_launch_fails_closed_on_a_live_pid_it_cannot_describe(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-    short_socket_path: Path,
-    caplog,
-) -> None:
-    """A live pid that cannot be excluded is not the same thing as a gone one.
-
-    Its record used to be retired on any unreadable identity, so a replacement
-    sidecar started beside a process that may still have been serving the socket.
-    No later attempt can clear this by itself, so the log has to name the pid that
-    is blocking the launch and the record file that points at it -- ``last_error``
-    alone only ever says ``memory_sidecar_unavailable``.
-    """
-
-    process = _orphan_process(tmp_path, stop_timeout_seconds=0.1, socket_path=short_socket_path)
-    record_path = _write_orphan_record(process, _orphan_record(process))
-    signalled: list[object] = []
-    spawns: list[tuple] = []
-
-    async def spawn(*args, **_kwargs):
-        spawns.append(args)
-        raise OSError("stop before a real sidecar is spawned")
-
-    # Our uid and the recorded creation time, but the cmdline is withheld: nothing
-    # observable rules this pid out as the sidecar the record names.
-    monkeypatch.setattr(
-        memory_process.psutil,
-        "Process",
-        _guarded_process_class(uid=os.getuid() if hasattr(os, "getuid") else None),
-    )
-    monkeypatch.setattr(memory_process, "_signal_owned_processes", lambda *_args: signalled.append(object()))
-    monkeypatch.setattr("core.memory.process.asyncio.create_subprocess_exec", spawn)
-
-    with caplog.at_level(logging.WARNING, logger=memory_process.logger.name):
-        assert asyncio.run(process.start()) is False
-
-    assert process.last_error == "memory_sidecar_unavailable"
-    assert signalled == []
-    assert spawns == []
-    assert record_path.exists()
-    assert "recorded sidecar identity could not be verified" in caplog.text
-    assert str(_ORPHAN_PID) in caplog.text
-    assert str(record_path) in caplog.text
-    # `logger.exception`, so the traceback reaches the log too.
-    assert "Traceback (most recent call last)" in caplog.text
-
-
-def test_sidecar_launch_proceeds_past_a_recycled_pid_owned_by_another_user(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-    short_socket_path: Path,
-) -> None:
-    """Failing closed must not turn into a permanent brick.
-
-    A pid recycled by another user's process is provably not our sidecar even when
-    that process withholds its cmdline, so the record is retired and the launch
-    continues rather than requiring manual intervention.
-    """
-
-    process = _orphan_process(tmp_path, stop_timeout_seconds=0.1, socket_path=short_socket_path)
-    record_path = _write_orphan_record(process, _orphan_record(process))
-    signalled: list[object] = []
-    spawns: list[tuple] = []
-    foreign_uid = (os.getuid() if hasattr(os, "getuid") else 0) + 1
-
-    async def spawn(*args, **_kwargs):
-        spawns.append(args)
-        raise OSError("stop before a real sidecar is spawned")
-
-    # The recorded creation time still matches, so only the foreign uid rules this
-    # pid out — exactly the fact macOS and Linux both disclose for any process.
-    monkeypatch.setattr(memory_process.psutil, "Process", _guarded_process_class(uid=foreign_uid))
-    monkeypatch.setattr(memory_process, "_signal_owned_processes", lambda *_args: signalled.append(object()))
-    monkeypatch.setattr("core.memory.process.asyncio.create_subprocess_exec", spawn)
-
-    assert asyncio.run(process.start()) is False
-    assert signalled == []
-    # The launch reached the spawn, so an unreadable stranger cannot wedge startup.
-    assert spawns
-    assert not record_path.exists()
-
-
-def test_sidecar_records_a_verified_launch_identity_privately(tmp_path: Path) -> None:
-    """The record must be owner-only, and an unverifiable identity is not recorded."""
-
-    process = _orphan_process(tmp_path)
-    process._ownership.record_path.parent.mkdir(parents=True, exist_ok=True)
-
-    process._ownership.record_launch(_ORPHAN_PID, _ORPHAN_CREATE_TIME, _ORPHAN_PID)
-    recorded = json.loads(process._ownership.record_path.read_text(encoding="utf-8"))
-
-    assert recorded == _orphan_record(process)
-    assert stat.S_IMODE(process._ownership.record_path.lstat().st_mode) == 0o600
-
-    process._ownership.record_path.unlink()
-    # An AccessDenied group member carries a negative sentinel instead of a
-    # creation time. Recording it would produce a record nothing can match, and
-    # skipping the write would launch a child no later boot can identify, so the
-    # launch has to fail instead.
-    with pytest.raises(RuntimeError, match="could not verify the sidecar creation time"):
-        process._ownership.record_launch(_ORPHAN_PID, -1.0, _ORPHAN_PID)
-
-    assert not process._ownership.record_path.exists()
-
-
-def _group_member_process_class(disclosures: dict[int, dict]):
-    """A ``psutil.Process`` stand-in for the members of a recorded process group.
-
-    Each entry lists what the OS discloses about that pid; anything left out is
-    withheld the way a real refusal is, and an unlisted pid is gone.
-    """
-
-    class _Member:
-        def __init__(self, process_id: int) -> None:
-            if process_id not in disclosures:
-                raise psutil.NoSuchProcess(pid=process_id)
-            self.pid = process_id
-            self._facts = disclosures[process_id]
-
-        def _disclosed(self, name: str):
-            value = self._facts.get(name)
-            if value is None:
-                raise psutil.AccessDenied(pid=self.pid)
-            return value
-
-        def status(self) -> str:
-            return psutil.STATUS_SLEEPING
-
-        def create_time(self) -> float:
-            return float(self._disclosed("create_time"))
-
-        def cmdline(self) -> list[str]:
-            return list(self._disclosed("cmdline"))
-
-        def environ(self) -> dict[str, str]:
-            return dict(self._disclosed("environ"))
-
-        def uids(self):
-            uid = self._disclosed("uid")
-            return SimpleNamespace(real=uid, effective=uid, saved=uid)
-
-    return _Member
-
-
-def _own_uid() -> int:
-    return os.getuid() if hasattr(os, "getuid") else 0
-
-
-def _recorded_group_disclosures(process: EverOSProcess) -> dict[int, dict]:
-    """Four live members of a dead leader's group: two ours, two to leave alone."""
-
-    return {
-        # Re-exec'd sidecar entrypoint: its command line names our socket.
-        _ORPHAN_DESCENDANT_PID: {
-            "create_time": _ORPHAN_CREATE_TIME + 1,
-            "uid": _own_uid(),
-            "cmdline": (sys.executable, "-m", _SIDECAR_ENTRYPOINT_MODULE, "--uds", str(process.socket_path)),
-        },
-        # A helper EverOS spawned: nothing in its command line, but it inherited
-        # the provider root from the environment the launch handed the sidecar.
-        _ORPHAN_GROUP_HELPER_PID: {
-            "create_time": _ORPHAN_CREATE_TIME + 2,
-            "uid": _own_uid(),
-            "environ": {"EVEROS_ROOT": str(process.provider_root), "HOME": "/tmp/child-home"},
-        },
-        # Same user, but nothing observable ties it to this installation.
-        _FOREIGN_GROUP_PID: {
-            "create_time": _ORPHAN_CREATE_TIME + 3,
-            "uid": _own_uid(),
-            "cmdline": ("/bin/sleep", "600"),
-            "environ": {"HOME": "/Users/someone"},
-        },
-        # Another user's process, which also withholds everything else.
-        _FOREIGN_UID_GROUP_PID: {
-            "create_time": _ORPHAN_CREATE_TIME + 4,
-            "uid": _own_uid() + 1,
-        },
-    }
-
-
-def test_sidecar_launch_reaps_group_members_a_gone_leader_left_behind(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-    caplog,
-) -> None:
-    """A leader that already exited still leaves its helpers holding the root.
-
-    The reap only ran for a leader classified ``OURS``, which requires it to be
-    alive. Once it had exited the record was deleted with no scan at all, so
-    same-group helpers kept the provider root open while a replacement sidecar
-    started -- the overlap this reap exists to prevent, reached the other way.
-
-    The gone leader can no longer vouch for the group, so each member must tie
-    itself to this installation. Members that cannot are logged and left running:
-    they may belong to an unrelated process that took the recorded pid and led a
-    group of the same number.
-    """
-
-    process = _orphan_process(tmp_path, stop_timeout_seconds=0.1)
-    record_path = _write_orphan_record(process, _orphan_record(process))
-    disclosures = _recorded_group_disclosures(process)
-    members = {pid: facts["create_time"] for pid, facts in disclosures.items()}
-    live = dict(members)
-    scanned_groups: list[int] = []
-    group_signals: list[tuple[int, int]] = []
-    signalled: list[dict[int, float]] = []
-
-    def snapshot_group(group: int) -> dict[int, float]:
-        scanned_groups.append(group)
-        return dict(members)
-
-    def signal_processes(identities, signum) -> None:
-        del signum
-        signalled.append(dict(identities))
-        for process_id in identities:
-            live.pop(process_id, None)
-
-    # The recorded leader is confirmed gone; only its group is left to work from.
-    monkeypatch.setattr(memory_process, "_inspect_process_identity", lambda _pid: None)
-    monkeypatch.setattr(memory_process, "_snapshot_process_group", snapshot_group)
-    monkeypatch.setattr(memory_process.psutil, "Process", _group_member_process_class(disclosures))
-    monkeypatch.setattr(memory_process.os, "killpg", lambda group, signum: group_signals.append((group, signum)))
-    monkeypatch.setattr(memory_process, "_signal_owned_processes", signal_processes)
-    monkeypatch.setattr(
-        memory_process,
-        "_live_owned_processes",
-        lambda identities: {pid: live[pid] for pid in identities if pid in live},
-    )
-
-    with caplog.at_level(logging.WARNING, logger=memory_process.logger.name):
-        asyncio.run(process._ownership.reap())
-
-    # Discovery only ever looked at the group the record names.
-    assert scanned_groups and set(scanned_groups) == {_ORPHAN_PID}
-    # One SIGTERM round, carrying only the two members that identified themselves.
-    assert signalled == [
-        {
-            _ORPHAN_DESCENDANT_PID: _ORPHAN_CREATE_TIME + 1,
-            _ORPHAN_GROUP_HELPER_PID: _ORPHAN_CREATE_TIME + 2,
-        }
-    ]
-    # A group holding members Avibe cannot claim is never signalled group-wide.
-    assert group_signals == []
-    assert str(_FOREIGN_GROUP_PID) in caplog.text
-    assert str(_FOREIGN_UID_GROUP_PID) in caplog.text
-    # The unclaimed members are still running, and must not wedge startup.
-    assert live == {
-        _FOREIGN_GROUP_PID: _ORPHAN_CREATE_TIME + 3,
-        _FOREIGN_UID_GROUP_PID: _ORPHAN_CREATE_TIME + 4,
-    }
-    assert not record_path.exists()
-
-
-def test_sidecar_launch_fails_closed_when_a_recorded_group_will_not_exit(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-    short_socket_path: Path,
-    caplog,
-) -> None:
-    """A surviving helper fails the launch, exactly as an unreapable orphan does.
-
-    Nothing later can clear this by itself, so the log has to name the group and the
-    record that points at it -- ``last_error`` only ever says
-    ``memory_sidecar_unavailable``.
-    """
-
-    process = _orphan_process(tmp_path, stop_timeout_seconds=0.1, socket_path=short_socket_path)
-    record_path = _write_orphan_record(process, _orphan_record(process))
-    disclosures = {
-        _ORPHAN_GROUP_HELPER_PID: {
-            "create_time": _ORPHAN_CREATE_TIME + 2,
-            "uid": _own_uid(),
-            "environ": {"EVEROS_ROOT": str(process.provider_root)},
-        }
-    }
-    members = {_ORPHAN_GROUP_HELPER_PID: _ORPHAN_CREATE_TIME + 2}
-    group_signals: list[tuple[int, int]] = []
-    signalled: list[int] = []
-    spawns: list[tuple] = []
-    kill_signal = getattr(signal, "SIGKILL", signal.SIGTERM)
-
-    async def spawn(*args, **_kwargs):
-        spawns.append(args)
-        raise OSError("stop before a real sidecar is spawned")
-
-    monkeypatch.setattr(memory_process, "_inspect_process_identity", lambda _pid: None)
-    monkeypatch.setattr(memory_process, "_snapshot_process_group", lambda _group: dict(members))
-    monkeypatch.setattr(memory_process.psutil, "Process", _group_member_process_class(disclosures))
-    monkeypatch.setattr(memory_process.os, "killpg", lambda group, signum: group_signals.append((group, signum)))
-    # The helper ignores every signal, so nothing proves the provider root is free.
-    monkeypatch.setattr(memory_process, "_signal_owned_processes", lambda _identities, signum: signalled.append(signum))
-    monkeypatch.setattr(memory_process, "_live_owned_processes", lambda identities: dict(identities))
-    monkeypatch.setattr("core.memory.process.asyncio.create_subprocess_exec", spawn)
-
-    with caplog.at_level(logging.WARNING, logger=memory_process.logger.name):
-        assert asyncio.run(process.start()) is False
-
-    assert process.last_error == "memory_sidecar_unavailable"
-    assert spawns == []
-    assert record_path.exists()
-    # Every member of this group identified itself, so the group-wide signal is
-    # allowed here -- unlike the mixed group above.
-    assert group_signals == [(_ORPHAN_PID, signal.SIGTERM), (_ORPHAN_PID, kill_signal)]
-    assert signalled == [signal.SIGTERM, kill_signal]
-    assert "orphaned sidecar group did not exit" in caplog.text
-    assert str(_ORPHAN_PID) in caplog.text
-    assert str(record_path) in caplog.text
-
-
-def test_sidecar_launch_proceeds_when_a_gone_leader_left_an_empty_group(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-    short_socket_path: Path,
-) -> None:
-    """The common case: the whole tree died with the service, so only the record is left."""
-
-    process = _orphan_process(tmp_path, stop_timeout_seconds=0.1, socket_path=short_socket_path)
-    record_path = _write_orphan_record(process, _orphan_record(process))
-    signalled: list[object] = []
-    spawns: list[tuple] = []
-
-    async def spawn(*args, **_kwargs):
-        spawns.append(args)
-        raise OSError("stop before a real sidecar is spawned")
-
-    monkeypatch.setattr(memory_process, "_inspect_process_identity", lambda _pid: None)
-    monkeypatch.setattr(memory_process, "_snapshot_process_group", lambda _group: {})
-    monkeypatch.setattr(memory_process, "_signal_owned_processes", lambda *_args: signalled.append(object()))
-    monkeypatch.setattr("core.memory.process.asyncio.create_subprocess_exec", spawn)
-
-    assert asyncio.run(process.start()) is False
-
-    assert signalled == []
-    # The launch got past the orphan check rather than failing closed on nothing.
-    assert spawns
-    assert not record_path.exists()
-
-
-@pytest.mark.parametrize("unscannable", ["record_from_an_older_build", "avibes_own_process_group"])
-def test_sidecar_launch_never_scans_a_group_it_must_not_signal(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-    unscannable: str,
-) -> None:
-    """Two records whose group must not be swept, for opposite reasons.
-
-    A build that predates the group field leaves nothing but a dead leader's pid,
-    which identifies no one. A record naming Avibe's own group cannot have been
-    written by ``_isolated_process_group``, and sweeping it would signal Avibe.
-    """
-
-    process = _orphan_process(tmp_path, stop_timeout_seconds=0.1)
-    record = _orphan_record(process)
-    if unscannable == "record_from_an_older_build":
-        del record["process_group"]
-    else:
-        record["process_group"] = os.getpgrp()
-    record_path = _write_orphan_record(process, record)
-    scanned_groups: list[int] = []
-    signalled: list[object] = []
-
-    def snapshot_group(group: int) -> dict[int, float]:
-        scanned_groups.append(group)
-        return {}
-
-    monkeypatch.setattr(memory_process, "_inspect_process_identity", lambda _pid: None)
-    monkeypatch.setattr(memory_process, "_snapshot_process_group", snapshot_group)
-    monkeypatch.setattr(memory_process, "_signal_owned_processes", lambda *_args: signalled.append(object()))
-    monkeypatch.setattr(memory_process.os, "killpg", lambda *_args: signalled.append(object()))
-
-    asyncio.run(process._ownership.reap())
-
-    assert scanned_groups == []
-    assert signalled == []
-    assert not record_path.exists()
-
-
-async def _succeed(*_args, **_kwargs) -> None:
-    return None
-
-
-def test_sidecar_start_fails_when_ownership_cannot_be_persisted(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-    short_socket_path: Path,
-) -> None:
-    """An unrecordable launch is a failed launch, and its child is reaped.
-
-    A swallowed write failure left a running sidecar that no record pointed at, so
-    a later crash produced an orphan the next boot could not see -- and that boot
-    started a replacement on the same provider root. ``_start_locked`` already
-    fails when in-memory ownership cannot be established; persisted ownership
-    follows the same rule, which also hands the child to its cleanup path.
-    """
-
-    class _Child:
-        pid = 999_999
-        returncode = None
-
-        async def wait(self) -> None:
-            return None
-
-        def send_signal(self, _signum) -> None:
-            return None
-
-    launches: list[_Child] = []
-    reaped: list[_Child] = []
-
-    async def spawn(*_args, **_kwargs) -> _Child:
-        child = _Child()
-        launches.append(child)
-        return child
-
-    process = _orphan_process(tmp_path, stop_timeout_seconds=0.1, socket_path=short_socket_path)
-    write_private_text = memory_process._write_private_text
-
-    def refuse_the_record(path: Path, contents: str) -> None:
-        if path == process._ownership.record_path:
-            raise OSError("record could not be written")
-        write_private_text(path, contents)
-
-    terminate_owned_tree = process._terminate_owned_tree
-
-    async def terminate(child, **kwargs) -> None:
-        reaped.append(child)
-        await terminate_owned_tree(child, **kwargs)
-
-    monkeypatch.setattr("core.memory.process.asyncio.create_subprocess_exec", spawn)
-    monkeypatch.setattr(memory_process, "_write_private_text", refuse_the_record)
-    monkeypatch.setattr(memory_process, "_snapshot_owned_processes", lambda pid, _group: {pid: _ORPHAN_CREATE_TIME})
-    monkeypatch.setattr(memory_process, "_owned_process_identity_is_live", lambda *_args: True)
-    monkeypatch.setattr(memory_process, "_live_owned_processes", lambda _identities: {})
-    monkeypatch.setattr(process, "_terminate_owned_tree", terminate)
-    # Everything after the record succeeds, so the unwritten record is the only
-    # thing that can fail this launch. Without these the start would fail on the
-    # absent socket instead, and the test would hold whether or not the record
-    # failure is respected.
-    monkeypatch.setattr(process, "_wait_for_ready", _succeed)
-    monkeypatch.setattr(process, "_secure_socket", lambda: None)
-    monkeypatch.setattr(process, "_assert_no_tcp_listener", lambda *_args, **_kwargs: None)
-
-    assert asyncio.run(process.start()) is False
-
-    assert process.last_error == "memory_sidecar_unavailable"
-    # The child was already tracked when the record failed, so the start failure
-    # reaped the tree it had just spawned instead of leaking it.
-    assert len(launches) == 1
-    assert reaped == launches
-    assert process._process is None
-    assert not process._ownership.record_path.exists()
-
-
-_UNUSABLE_RECORDS: dict[str, bytes] = {
-    "truncated": b'{"pid": 424242, "create_ti',
-    "not_json": b"\x00\x01 not a record at all",
-    "oversized": b"{}" + b" " * (5 * 1024),
-    "no_pid": b'{"create_time": 1700000000.5}',
-}
-
-
-def _sidecar_scan_disclosures(process: EverOSProcess) -> dict[int, dict]:
-    """One process that is our sidecar, and three that only look like it."""
-
-    return {
-        # Our entrypoint, serving our socket: the anchor.
-        _ORPHAN_PID: {
-            "create_time": _ORPHAN_CREATE_TIME,
-            "uid": _own_uid(),
-            "cmdline": (sys.executable, "-m", _SIDECAR_ENTRYPOINT_MODULE, "--uds", str(process.socket_path)),
-            "environ": {"EVEROS_ROOT": str(process.provider_root)},
-        },
-        # A helper in the anchor's group, claimed through group membership.
-        _ORPHAN_GROUP_HELPER_PID: {
-            "create_time": _ORPHAN_CREATE_TIME + 2,
-            "uid": _own_uid(),
-            "environ": {"EVEROS_ROOT": str(process.provider_root)},
-        },
-        # The short-lived processing probe. It carries the same environment, which
-        # is exactly why the machine-wide test may not accept an environment match.
-        _FOREIGN_GROUP_PID: {
-            "create_time": _ORPHAN_CREATE_TIME + 3,
-            "uid": _own_uid(),
-            "cmdline": (sys.executable, "-m", _SIDECAR_ENTRYPOINT_MODULE, "--probe-processing"),
-            "environ": {"EVEROS_ROOT": str(process.provider_root)},
-        },
-        # Somebody looking at the provider root. Naming a path is not owning it.
-        _FOREIGN_UID_GROUP_PID: {
-            "create_time": _ORPHAN_CREATE_TIME + 4,
-            "uid": _own_uid(),
-            "cmdline": ("/bin/ls", "-l", str(process.provider_root)),
-        },
-    }
-
-
-def test_sidecar_launch_reaps_a_live_sidecar_an_unusable_record_cannot_name(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-    caplog,
-) -> None:
-    """A record that exists but cannot be parsed is not the same as no record.
-
-    Both used to read as ``None`` and retire the file, so a truncated or unreadable
-    record discarded the only ownership evidence and let a replacement launch
-    against a socket and provider root the previous run's sidecar may still have
-    been holding. Ownership is rebuilt from live processes instead, which needs no
-    record: our entrypoint serving our socket, plus that anchor's own group.
-
-    Failing closed on the corrupt file instead would have been unrecoverable --
-    nothing repairs it, so every later start would fail forever.
-    """
-
-    process = _orphan_process(tmp_path, stop_timeout_seconds=0.1)
-    record_path = process._ownership.record_path
-    record_path.parent.mkdir(parents=True, exist_ok=True)
-    record_path.write_bytes(_UNUSABLE_RECORDS["truncated"])
-    disclosures = _sidecar_scan_disclosures(process)
-    stub = _group_member_process_class(disclosures)
-    group_members = {
-        _ORPHAN_PID: _ORPHAN_CREATE_TIME,
-        _ORPHAN_GROUP_HELPER_PID: _ORPHAN_CREATE_TIME + 2,
-    }
-    live = dict(group_members)
-    group_signals: list[tuple[int, int]] = []
-    signalled: list[dict[int, float]] = []
-
-    def signal_processes(identities, signum) -> None:
-        del signum
-        signalled.append(dict(identities))
-        for pid in identities:
-            live.pop(pid, None)
-
-    monkeypatch.setattr(memory_process.psutil, "Process", stub)
-    monkeypatch.setattr(memory_process.psutil, "process_iter", lambda: [stub(pid) for pid in disclosures])
-    monkeypatch.setattr(memory_process.os, "killpg", lambda group, signum: group_signals.append((group, signum)))
-    # The sidecar leads a session of its own, so its helpers share its group.
-    monkeypatch.setattr(memory_process, "_isolated_process_group", lambda pid: pid)
-    monkeypatch.setattr(memory_process, "_snapshot_process_group", lambda _group: dict(group_members))
-    monkeypatch.setattr(memory_process, "_signal_owned_processes", signal_processes)
-    monkeypatch.setattr(
-        memory_process,
-        "_live_owned_processes",
-        lambda identities: {pid: live[pid] for pid in identities if pid in live},
-    )
-
-    with caplog.at_level(logging.WARNING, logger=memory_process.logger.name):
-        asyncio.run(process._ownership.reap())
-
-    # The anchor and its group helper, and neither look-alike.
-    assert signalled == [group_members]
-    # Every member of the anchor's group is claimed, so the group signal is allowed.
-    assert group_signals == [(_ORPHAN_PID, signal.SIGTERM)]
-    assert str(_FOREIGN_GROUP_PID) not in caplog.text
-    assert str(_FOREIGN_UID_GROUP_PID) not in caplog.text
-    assert not record_path.exists()
-
-
-@pytest.mark.parametrize("corruption", sorted(_UNUSABLE_RECORDS))
-def test_sidecar_launch_proceeds_when_an_unusable_record_names_nothing_running(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-    short_socket_path: Path,
-    corruption: str,
-) -> None:
-    """An unusable record must not brick startup when nothing of ours survives."""
-
-    process = _orphan_process(tmp_path, stop_timeout_seconds=0.1, socket_path=short_socket_path)
-    record_path = process._ownership.record_path
-    record_path.parent.mkdir(parents=True, exist_ok=True)
-    record_path.write_bytes(_UNUSABLE_RECORDS[corruption])
-    disclosures = _sidecar_scan_disclosures(process)
-    # Only the look-alikes are running; the sidecar itself is gone.
-    running = {pid: facts for pid, facts in disclosures.items() if pid != _ORPHAN_PID}
-    stub = _group_member_process_class(running)
-    signalled: list[object] = []
-    spawns: list[tuple] = []
-
-    async def spawn(*args, **_kwargs):
-        spawns.append(args)
-        raise OSError("stop before a real sidecar is spawned")
-
-    monkeypatch.setattr(memory_process.psutil, "Process", stub)
-    monkeypatch.setattr(memory_process.psutil, "process_iter", lambda: [stub(pid) for pid in running])
-    monkeypatch.setattr(memory_process, "_signal_owned_processes", lambda *_args: signalled.append(object()))
-    monkeypatch.setattr("core.memory.process.asyncio.create_subprocess_exec", spawn)
-
-    assert asyncio.run(process.start()) is False
-
-    assert signalled == []
-    # The launch got past the record check rather than failing closed on a file.
-    assert spawns
-    assert not record_path.exists()
-
-
-def test_sidecar_launch_scans_for_processes_only_when_a_record_exists(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-    short_socket_path: Path,
-) -> None:
-    """The ordinary first boot must not pay for a machine-wide process scan."""
-
-    process = _orphan_process(tmp_path, stop_timeout_seconds=0.1, socket_path=short_socket_path)
-    scans: list[object] = []
-    spawns: list[tuple] = []
-
-    async def spawn(*args, **_kwargs):
-        spawns.append(args)
-        raise OSError("stop before a real sidecar is spawned")
-
-    monkeypatch.setattr(memory_process.psutil, "process_iter", lambda: scans.append(object()) or [])
-    monkeypatch.setattr("core.memory.process.asyncio.create_subprocess_exec", spawn)
-
-    assert not process._ownership.record_path.exists()
-    assert asyncio.run(process.start()) is False
-
-    assert scans == []
-    assert spawns
-
-
-def test_sidecar_launch_fails_closed_when_an_unusable_record_names_a_live_sidecar(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-    short_socket_path: Path,
-    caplog,
-) -> None:
-    """A sidecar that will not exit fails the launch and keeps the record.
-
-    The record is unusable, so the log is the only thing that can point an operator
-    at what is blocking the start.
-    """
-
-    process = _orphan_process(tmp_path, stop_timeout_seconds=0.1, socket_path=short_socket_path)
-    record_path = process._ownership.record_path
-    record_path.parent.mkdir(parents=True, exist_ok=True)
-    record_path.write_bytes(_UNUSABLE_RECORDS["truncated"])
-    disclosures = {_ORPHAN_PID: _sidecar_scan_disclosures(process)[_ORPHAN_PID]}
-    stub = _group_member_process_class(disclosures)
-    spawns: list[tuple] = []
-
-    async def spawn(*args, **_kwargs):
-        spawns.append(args)
-        raise OSError("stop before a real sidecar is spawned")
-
-    monkeypatch.setattr(memory_process.psutil, "Process", stub)
-    monkeypatch.setattr(memory_process.psutil, "process_iter", lambda: [stub(_ORPHAN_PID)])
-    monkeypatch.setattr(memory_process, "_isolated_process_group", lambda _pid: None)
-    # The sidecar ignores every signal.
-    monkeypatch.setattr(memory_process, "_signal_owned_processes", lambda *_args: None)
-    monkeypatch.setattr(memory_process, "_live_owned_processes", lambda identities: dict(identities))
-    monkeypatch.setattr("core.memory.process.asyncio.create_subprocess_exec", spawn)
-
-    with caplog.at_level(logging.WARNING, logger=memory_process.logger.name):
-        assert asyncio.run(process.start()) is False
-
-    assert process.last_error == "memory_sidecar_unavailable"
-    assert spawns == []
-    assert record_path.exists()
-    assert "sidecar left by an unusable record did not exit" in caplog.text
-    assert str(_ORPHAN_PID) in caplog.text
-    assert str(record_path) in caplog.text
-
-
-class _ExitedChild:
-    """A direct child that has already exited, as ``_watch_child`` finds it."""
-
-    pid = _ORPHAN_PID
-    returncode = 0
-
-    async def wait(self) -> None:
-        return None
-
-    def send_signal(self, _signum) -> None:
-        raise AssertionError("a child that already exited must not be signalled")
-
-
-def _supervising(process: EverOSProcess, child: _ExitedChild) -> None:
-    """Put the supervisor in the state a running sidecar leaves behind."""
-
-    process._process = child
-    process._process_group = _ORPHAN_PID
-    process._owned_processes = {_ORPHAN_PID: _ORPHAN_CREATE_TIME}
-
-
-def _late_helper_disclosures(process: EverOSProcess) -> dict[int, dict]:
-    """One helper the sidecar spawned after the monitor's last snapshot.
-
-    The recorded leader is deliberately absent, so ``psutil`` answers
-    ``NoSuchProcess`` for it exactly as it does for a child that has exited.
-    """
-
-    return {
-        _ORPHAN_GROUP_HELPER_PID: {
-            "create_time": _ORPHAN_CREATE_TIME + 2,
-            "uid": _own_uid(),
-            "environ": {"EVEROS_ROOT": str(process.provider_root)},
-        }
-    }
-
-
-def test_sidecar_notifies_reaped_callback_only_after_tree_cleanup(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-) -> None:
-    """The runtime handoff begins after the exited child's tree is reaped."""
-
-    events: list[str] = []
-
-    async def on_reaped() -> None:
-        assert process._process is None
-        events.append("reaped")
-
-    process = EverOSProcess(
-        sys.executable,
-        effective_home=tmp_path,
-        settings=_settings(),
-        on_reaped=on_reaped,
-    )
-    child = _ExitedChild()
-    _supervising(process, child)
-    process._desired_running = False
-
-    async def terminate(*_args, **_kwargs) -> None:
-        events.append("tree-cleaned")
-
-    monkeypatch.setattr(process, "_terminate_owned_tree", terminate)
-    monkeypatch.setattr(process._ownership, "retire_if_group_is_clear", lambda _group: events.append("retired"))
-
-    asyncio.run(process._watch_child(child))
-
-    assert events == ["tree-cleaned", "retired", "reaped"]
-
-
-def test_sidecar_start_failure_after_host_handoff_notifies_reaped(tmp_path: Path) -> None:
-    """A pre-spawn launch failure leaves the host free to reclaim the call log."""
-
-    reaped = 0
-
-    async def on_reaped() -> None:
-        nonlocal reaped
-        reaped += 1
-
-    process = EverOSProcess(
-        sys.executable,
-        effective_home=tmp_path,
-        settings=_settings(),
-        on_reaped=on_reaped,
-    )
-    process._consecutive_failures = 4
-    process._validate_launch_inputs = lambda: (_ for _ in ()).throw(RuntimeError("launch rejected"))
-
-    assert asyncio.run(process.start()) is False
-    assert reaped == 1
-    assert process._restart_task is None
-
-
-async def test_sidecar_restart_releases_process_lock_while_host_handoff_waits(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-) -> None:
-    """Stop wins while a restart waits for the controller-owned call-log lock."""
-
-    callback_entered = asyncio.Event()
-    release_callback = asyncio.Event()
-    launches = 0
-
-    async def before_start() -> None:
-        callback_entered.set()
-        await release_callback.wait()
-
-    process = EverOSProcess(
-        sys.executable,
-        effective_home=tmp_path,
-        settings=_settings(),
-        before_start=before_start,
-    )
-    process._desired_running = True
-
-    async def start_locked() -> bool:
-        nonlocal launches
-        launches += 1
-        return True
-
-    monkeypatch.setattr(process, "_start_locked", start_locked)
-
-    restarting = asyncio.create_task(process._restart_after(0))
-    await asyncio.wait_for(callback_entered.wait(), timeout=1)
-
-    # This models a callback already in flight when Stop begins. It must
-    # not retain the process lock while waiting on controller ownership.
-    await asyncio.wait_for(process.stop(), timeout=1)
-    release_callback.set()
-    await asyncio.wait_for(restarting, timeout=1)
-    assert launches == 0
-
-
-def test_fake_sidecar_failed_start_notifies_reaped_for_runtime_handoff() -> None:
-    reaped = 0
-
-    async def on_reaped() -> None:
-        nonlocal reaped
-        reaped += 1
-
-    process = FakeEverOSProcess(
-        start_results=deque([False]),
-        on_reaped=on_reaped,
-    )
-
-    assert asyncio.run(process.start()) is False
-    assert reaped == 1
-
-
-@pytest.mark.parametrize("group_holds_a_survivor", [True, False])
-def test_sidecar_cleanup_retires_the_record_only_once_its_group_is_clear(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-    caplog,
-    group_holds_a_survivor: bool,
-) -> None:
-    """A reaped leader does not prove its group is empty.
-
-    When the sidecar spawns a helper after the monitor's last snapshot and then
-    exits, that helper is in none of the captured identities. Rediscovery is
-    anchored on a live leader, the group signal is refused because the unknown
-    member cannot be confirmed, and the wait then succeeds over the identities it
-    does hold. Retiring the record on that evidence threw away the next launch's
-    only route to the survivor -- the recorded group -- so a replacement sidecar
-    came up beside it on the same provider root.
-    """
-
-    process = _orphan_process(tmp_path, stop_timeout_seconds=0.1)
-    record_path = _write_orphan_record(process, _orphan_record(process))
-    disclosures = _late_helper_disclosures(process) if group_holds_a_survivor else {}
-    survivors = {_ORPHAN_GROUP_HELPER_PID: _ORPHAN_CREATE_TIME + 2} if group_holds_a_survivor else {}
-    group_signals: list[tuple[int, int]] = []
-    child = _ExitedChild()
-    _supervising(process, child)
-    process._desired_running = False
-
-    monkeypatch.setattr(memory_process.psutil, "Process", _group_member_process_class(disclosures))
-    monkeypatch.setattr(memory_process, "_snapshot_process_group", lambda _group: dict(survivors))
-    monkeypatch.setattr(memory_process.os, "killpg", lambda group, signum: group_signals.append((group, signum)))
-
-    with caplog.at_level(logging.WARNING, logger=memory_process.logger.name):
-        asyncio.run(process._watch_child(child))
-
-    # The cleanup itself is unchanged: it never signals a group holding a member
-    # it cannot confirm, and it finishes rather than failing.
-    assert group_signals == []
-    assert process._process is None
-    assert record_path.exists() is group_holds_a_survivor
-    if group_holds_a_survivor:
-        assert "Keeping the EverOS ownership record" in caplog.text
-        assert str(_ORPHAN_GROUP_HELPER_PID) in caplog.text
-        assert str(record_path) in caplog.text
-
-
-def test_sidecar_stop_keeps_the_record_while_its_group_holds_a_survivor(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-) -> None:
-    """Stop retires the record on the same evidence, so it needs the same guard.
-
-    Nothing sweeps here -- there is no launch to fail -- so the record is what
-    carries the survivor to the next one.
-    """
-
-    process = _orphan_process(tmp_path, stop_timeout_seconds=0.1)
-    record_path = _write_orphan_record(process, _orphan_record(process))
-    child = _ExitedChild()
-    _supervising(process, child)
-
-    monkeypatch.setattr(memory_process.psutil, "Process", _group_member_process_class(_late_helper_disclosures(process)))
-    monkeypatch.setattr(
-        memory_process,
-        "_snapshot_process_group",
-        lambda _group: {_ORPHAN_GROUP_HELPER_PID: _ORPHAN_CREATE_TIME + 2},
-    )
-    monkeypatch.setattr(memory_process.os, "killpg", lambda *_args: None)
-
-    asyncio.run(process.stop())
-
-    assert process._process is None
-    assert record_path.exists()
-
-
-def test_sidecar_orphan_reap_sweeps_the_group_before_retiring_the_record(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-) -> None:
-    """Reaping the recorded root leaves the same gap as reaping a live child.
-
-    ``_terminate_orphan_tree`` rediscovers only while the root is alive, so a
-    helper spawned in its last moments is proven by nothing. With the root now
-    gone, the leader-gone sweep is the follow-up, and a group it cannot clear
-    fails the launch instead of retiring the record and spawning beside it.
-    """
-
-    process = _orphan_process(tmp_path, stop_timeout_seconds=0.1)
-    record_path = _write_orphan_record(process, _orphan_record(process))
-    survivors = {_ORPHAN_GROUP_HELPER_PID: _ORPHAN_CREATE_TIME + 2}
-    group_signals: list[tuple[int, int]] = []
-
-    async def reaped(*_args, **_kwargs) -> bool:
-        return True
-
-    monkeypatch.setattr(
-        memory_process,
-        "_inspect_process_identity",
-        lambda _pid: _orphan_identity(process),
-    )
-    monkeypatch.setattr(process._ownership, "_terminate_orphan_tree", reaped)
-    monkeypatch.setattr(memory_process.psutil, "Process", _group_member_process_class(_late_helper_disclosures(process)))
-    monkeypatch.setattr(memory_process, "_snapshot_process_group", lambda _group: dict(survivors))
-    monkeypatch.setattr(memory_process.os, "killpg", lambda group, signum: group_signals.append((group, signum)))
-    # The survivor ignores every signal.
-    monkeypatch.setattr(memory_process, "_signal_owned_processes", lambda *_args: None)
-    monkeypatch.setattr(memory_process, "_live_owned_processes", lambda identities: dict(identities))
-
-    with pytest.raises(RuntimeError, match="orphaned sidecar group did not exit"):
-        asyncio.run(process._ownership.reap())
-
-    assert record_path.exists()
-    # Every member of that group is claimed, so the group signal is allowed here.
-    assert group_signals == [(_ORPHAN_PID, signal.SIGTERM), (_ORPHAN_PID, getattr(signal, "SIGKILL", signal.SIGTERM))]
+        await runtime.close()
+    assert isinstance(result, CaptureAccepted)
+    pinned_files = tuple((expected_pin_root / "bundles").glob("*/*"))
+    assert len(pinned_files) == 1
+    assert pinned_files[0].read_bytes() == b"runtime-owned attachment"
+    assert not attachment_pin_root(global_home).exists()
